requirements.txt
folium==0.14.0
haversine==2.8.0
numpy==1.24.4
pandas==2.0.3
Pillow==9.5.0
plotly==5.16.1
streamlit==1.26.0
streamlit-folium==0.13.0
python==3.8.17